마지막으로 수정 완료git add * 12번은 제출 안할것임

+++ c02/ex10/ft_strlcpy.c
@@ -6,16 +6,16 @@
 /*   By: kjo <kjo@student.42.seoul.kr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/05 00:27:12 by kyeongbin         #+#    #+#             */
-/*   Updated: 2020/08/06 22:33:46 by kjo              ###   ########.fr       */
+/*   Updated: 2020/08/08 09:17:39 by kjo              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size)
 {
-	unsigned int i;
+ 	int i;
 
 	i = 0;
-	while (src[i] != '\0' && i < size - 1)
+	while (src[i] != '\0' && i <(int)size - 1)
 	{
 		dest[i] = src[i];
 		i++;
